Module zcash_primitives::consensus

Consensus logic and parameters.

Modules

testingtest or test-dependencies

Structs

BlockHeight

A wrapper type representing blockchain heights. Safe conversion from various integer types, as well as addition and subtraction, are provided.

MainNetwork

Marker struct for the production network.

TestNetwork

Marker struct for the test network.

Enums

BranchId

A globally-unique identifier for a set of consensus rules within the Zcash chain.

Network

NetworkUpgrade

An event that occurs at a specified height on the Zcash chain, at which point the consensus rules enforced by the network are altered.

Constants

H0

MAIN_NETWORK

TEST_NETWORK

ZIP212_GRACE_PERIOD

Traits

Parameters

Zcash consensus parameters.