zcash_primitives
Module legacy

Support for legacy transparent addresses and scripts.

Modules

keystransparent-inputs

Transparent key components.

testingtest-dependencies

Structs

Script

A serialized script, used inside transparent inputs and outputs of a transaction.

Enums

TransparentAddress

A transparent address corresponding to either a public key hash or a script hash.