Trait zcash_primitives::legacy::keys::IncomingViewingKey

pub trait IncomingViewingKey: SealedChangeLevelKey + Sized {
    fn derive_address(
        &self, 
        child_index: u32
    ) -> Result<TransparentAddress, Error> { ... }
    fn default_address(&self) -> (TransparentAddress, u32) { ... }
    fn serialize(&self) -> Vec<u8> { ... }
    fn deserialize(data: &[u8; 65]) -> Result<Self, Error> { ... }
}

Provided Methods

fn derive_address(&self, child_index: u32) -> Result<TransparentAddress, Error>

Derives a transparent address at the provided child index.

fn default_address(&self) -> (TransparentAddress, u32)

Searches the space of child indexes for an index that will generate a valid transparent address, and returns the resulting address and the index at which it was generated.

fn serialize(&self) -> Vec<u8>

fn deserialize(data: &[u8; 65]) -> Result<Self, Error>