Module zcash_primitives::memo

Structs for handling encrypted memos.

Structs

MemoBytes

The unencrypted memo bytes received alongside a shielded note in a Zcash transaction.

TextMemo

Type-safe wrapper around String to enforce memo length requirements.

Enums

Error

Errors that may result from attempting to construct an invalid memo.

Memo

An unencrypted memo received alongside a shielded note in a Zcash transaction.