Module zcash_primitives::merkle_tree

Implementation of a Merkle tree of commitments used to prove the existence of notes.

Modules

incremental

Implementations of serialization and parsing for Orchard note commitment trees.

testingtest or test-dependencies

Structs

CommitmentTree

A Merkle tree of note commitments.

IncrementalWitness

An updatable witness to a path from a position in a particular CommitmentTree.

MerklePath

A path from a position in a particular commitment tree to the root of that tree.

Traits

HashSer

A hashable node within a Merkle tree.

Hashable

A hashable node within a Merkle tree.