Module zcash_primitives::sapling

Structs and constants specific to the Sapling shielded pool.

Modules

group_hash

Implementation of group hashing into Jubjub.

keys

Sapling key components.

note_encryption

Implementation of in-band secret distribution for Zcash transactions.

pedersen_hash

Implementation of the Pedersen hash function used in Sapling.

prover

Abstractions over the proving system and parameters.

redjubjub

Implementation of RedJubjub, a specialization of RedDSA to the Jubjub curve.

testing

util

Structs

Diversifier

Node

A node within the Sapling commitment tree.

Note

NoteValue

Nullifier

Typesafe wrapper for nullifier values.

NullifierDerivingKey

A key used to derive the nullifier for a Sapling note.

PaymentAddress

A Sapling payment address.

ProofGenerationKey

SaplingIvk

ValueCommitment

ViewingKey

Enums

Rseed

Enum for note randomness before and after ZIP 212.

Constants

SAPLING_COMMITMENT_TREE_DEPTH

SAPLING_COMMITMENT_TREE_DEPTH_U8

Functions

merkle_hash

Compute a parent node in the Sapling commitment tree given its two children.

spend_sig

Create the spendAuthSig for a Sapling SpendDescription.