Function zcash_primitives::sapling::note_encryption::sapling_note_encryption

pub fn sapling_note_encryption<R: RngCore, P: Parameters>(
    ovk: Option<OutgoingViewingKey>,
    note: Note,
    to: PaymentAddress,
    memo: MemoBytes,
    rng: &mut R
) -> NoteEncryption<SaplingDomain<P>>

Creates a new encryption context for the given note.

Setting ovk to None represents the ovk = ⊥ case, where the note cannot be recovered by the sender.

NB: the example code here only covers the post-Canopy case.

Examples

use ff::Field;
use rand_core::OsRng;
use zcash_primitives::{
    keys::{OutgoingViewingKey, prf_expand},
    consensus::{TEST_NETWORK, TestNetwork, NetworkUpgrade, Parameters},
    memo::MemoBytes,
    sapling::{
        note_encryption::sapling_note_encryption,
        util::generate_random_rseed,
        Diversifier, PaymentAddress, Rseed, ValueCommitment
    },
};

let mut rng = OsRng;

let diversifier = Diversifier([0; 11]);
let pk_d = diversifier.g_d().unwrap();
let to = PaymentAddress::from_parts(diversifier, pk_d).unwrap();
let ovk = Some(OutgoingViewingKey([0; 32]));

let value = 1000;
let rcv = jubjub::Fr::random(&mut rng);
let cv = ValueCommitment {
    value,
    randomness: rcv.clone(),
};
let height = TEST_NETWORK.activation_height(NetworkUpgrade::Canopy).unwrap();
let rseed = generate_random_rseed(&TEST_NETWORK, height, &mut rng);
let note = to.create_note(value, rseed).unwrap();
let cmu = note.cmu();

let mut enc = sapling_note_encryption::<_, TestNetwork>(ovk, note, to, MemoBytes::empty(), &mut rng);
let encCiphertext = enc.encrypt_note_plaintext();
let outCiphertext = enc.encrypt_outgoing_plaintext(&cv.commitment().into(), &cmu, &mut rng);