Function zcash_primitives::sapling::note_encryption::try_sapling_output_recovery

pub fn try_sapling_output_recovery<P: Parameters>(
    params: &P, 
    height: BlockHeight, 
    ovk: &OutgoingViewingKey, 
    output: &OutputDescription<GrothProofBytes>
) -> Option<(Note, PaymentAddress, MemoBytes)>

Recovery of the full note plaintext by the sender.

Attempts to decrypt and validate the given enc_ciphertext using the given ovk. If successful, the corresponding Sapling note and memo are returned, along with the PaymentAddress to which the note was sent.

Implements section 4.19.3 of the Zcash Protocol Specification.