Struct zcash_primitives::transaction::builder::Builder

pub struct Builder<'a, P, R> { /* private fields */ }

Generates a Transaction from its inputs and outputs.

Implementations

impl<'a, P: Parameters, R: RngCore> Builder<'a, P, R>

pub fn test_only_new_with_rng(
params: P,
height: BlockHeight,
rng: R
) -> Builder<'a, P, R>

Creates a new Builder targeted for inclusion in the block with the given height and randomness source, using default values for general transaction fields.

Default values

The expiry height will be set to the given height plus the default transaction expiry delta (20 blocks).

WARNING: DO NOT USE IN PRODUCTION

pub fn mock_build(
self
) -> Result<(Transaction, SaplingMetadata), Error<Infallible>>

impl<'a, P, R> Builder<'a, P, R>

pub fn params(&self) -> &P

Returns the network parameters that the builder has been configured for.

pub fn target_height(&self) -> BlockHeight

Returns the target height of the transaction under construction.

pub fn transparent_inputs(&self) -> &[impl InputView]

Returns the set of transparent inputs currently committed to be consumed by the transaction.

pub fn transparent_outputs(&self) -> &[impl OutputView]

Returns the set of transparent outputs currently set to be produced by the transaction.

pub fn sapling_inputs(&self) -> &[impl InputView<()>]

Returns the set of Sapling inputs currently committed to be consumed by the transaction.

pub fn sapling_outputs(&self) -> &[impl OutputView]

Returns the set of Sapling outputs currently set to be produced by the transaction.

impl<'a, P: Parameters> Builder<'a, P, OsRng>

pub fn new(params: P, target_height: BlockHeight) -> Self

Creates a new Builder targeted for inclusion in the block with the given height, using default values for general transaction fields and the default OS random.

Default values

The expiry height will be set to the given height plus the default transaction expiry delta (20 blocks).

impl<'a, P: Parameters, R: RngCore + CryptoRng> Builder<'a, P, R>

pub fn new_with_rng(
params: P,
target_height: BlockHeight,
rng: R
) -> Builder<'a, P, R>

Creates a new Builder targeted for inclusion in the block with the given height and randomness source, using default values for general transaction fields.

Default values

The expiry height will be set to the given height plus the default transaction expiry delta (20 blocks).

impl<'a, P: Parameters, R: RngCore> Builder<'a, P, R>

pub fn add_sapling_spend(
&mut self,
extsk: ExtendedSpendingKey,
diversifier: Diversifier,
note: Note,
merkle_path: MerklePath<Node>
) -> Result<(), Error>

Adds a Sapling note to be spent in this transaction.

Returns an error if the given Merkle path does not have the same anchor as the paths for previous Sapling notes.

pub fn add_sapling_output(
&mut self,
ovk: Option<OutgoingViewingKey>,
to: PaymentAddress,
value: Amount,
memo: MemoBytes
) -> Result<(), Error>

Adds a Sapling address to send funds to.

pub fn add_transparent_input(
&mut self,
sk: SecretKey,
utxo: OutPoint,
coin: TxOut
) -> Result<(), Error>

Available on crate feature transparent-inputs only.

Adds a transparent coin to be spent in this transaction.

pub fn add_transparent_output(
&mut self,
to: &TransparentAddress,
value: Amount
) -> Result<(), Error>

Adds a transparent address to send funds to.

pub fn with_progress_notifier(&mut self, progress_notifier: Sender<Progress>)

Sets the notifier channel, where progress of building the transaction is sent.

An update is sent after every Spend or Output is computed, and the u32 sent represents the total steps completed so far. It will eventually send number of spends + outputs. If there’s an error building the transaction, the channel is closed.

pub fn build<FR: FeeRule>(
self,
prover: &impl TxProver,
fee_rule: &FR
) -> Result<(Transaction, SaplingMetadata), Error<FR::Error>>

Builds a transaction from the configured spends and outputs.

Upon success, returns a tuple containing the final transaction, and the SaplingMetadata generated during the build process.

pub fn build_zfuture<FR: FutureFeeRule>(
self,
prover: &impl TxProver,
fee_rule: &FR
) -> Result<(Transaction, SaplingMetadata), Error<FR::Error>>

Builds a transaction from the configured spends and outputs.

Upon success, returns a tuple containing the final transaction, and the SaplingMetadata generated during the build process.