zcash_primitives::transaction::components
Module transparent

Structs representing the components within Zcash transactions.

Modules

builder

Types and functions for building transparent transaction components.

pczt

PCZT support for transparent Zcash.

testingtest-dependencies

Structs

Authorized

Bundle

EffectsOnly

Marker type for a bundle that contains no authorizing data, and the necessary input information for creating sighashes.

OutPoint

TxIn

TxOut

Traits

Authorization

MapAuth