Trait zcash_primitives::transaction::fees::FutureFeeRule

pub trait FutureFeeRule: FeeRule {
    fn fee_required_zfuture<P: Parameters>(
        &self,
        params: &P,
        target_height: BlockHeight,
        transparent_inputs: &[impl InputView],
        transparent_outputs: &[impl OutputView],
        sapling_input_count: usize,
        sapling_output_count: usize,
        tze_inputs: &[impl InputView],
        tze_outputs: &[impl OutputView]
    ) -> Result<Amount, Self::Error>;
}

A trait that represents the ability to compute the fees that must be paid by a transaction having a specified set of inputs and outputs, for use when experimenting with the TZE feature.

Required Methods

fn fee_required_zfuture<P: Parameters>(
&self,
params: &P,
target_height: BlockHeight,
transparent_inputs: &[impl InputView],
transparent_outputs: &[impl OutputView],
sapling_input_count: usize,
sapling_output_count: usize,
tze_inputs: &[impl InputView],
tze_outputs: &[impl OutputView]
) -> Result<Amount, Self::Error>

Computes the total fee required for a transaction given the provided inputs and outputs.

Implementations of this method should compute the fee amount given exactly the inputs and outputs specified, and should NOT compute speculative fees given any additional change outputs that may need to be created in order for inputs and outputs to balance.