zcash_primitives::transaction::fees::transparent

Trait InputView

pub trait InputView: Debug {
    // Required methods
    fn outpoint(&self) -> &OutPoint;
    fn coin(&self) -> &TxOut;

    // Provided method
    fn serialized_size(&self) -> InputSize { ... }
}

This trait provides a minimized view of a transparent input suitable for use in fee and change computation.

Required Methods

fn outpoint(&self) -> &OutPoint

The outpoint to which the input refers.

fn coin(&self) -> &TxOut

The previous output being spent.

Provided Methods

fn serialized_size(&self) -> InputSize

The size of the transparent script required to spend this input.