zcash_primitives::transaction::fees::transparent

Trait OutputView

pub trait OutputView: Debug {
    // Required methods
    fn value(&self) -> Zatoshis;
    fn script_pubkey(&self) -> &Script;

    // Provided method
    fn serialized_size(&self) -> usize { ... }
}

This trait provides a minimized view of a transparent output suitable for use in fee and change computation.

Required Methods

fn value(&self) -> Zatoshis

Returns the value of the output being created.

fn script_pubkey(&self) -> &Script

Returns the script corresponding to the newly created output.

Provided Methods

fn serialized_size(&self) -> usize

Returns the serialized size of the txout.