Function zcash_primitives::transaction::sighash::signature_hash

pub fn signature_hash<'a, TA: TransparentAuthorizingContext, SA: Authorization<Proof = GrothProofBytes>, A: Authorization<SaplingAuth = SA, TransparentAuth = TA>>(
    tx: &TransactionData<A>, 
    signable_input: &SignableInput<'a>, 
    txid_parts: &TxDigests<Blake2bHash>
) -> SignatureHash

Computes the signature hash for an input to a transaction, given the full data of the transaction, the input being signed, and the set of precomputed hashes produced in the construction of the transaction ID.