zcash_primitives

Module zip32

👎Deprecated: This module is deprecated; use the zip32 crate instead.

Modules

arbitrary

Arbitrary key derivation.

fingerprint

Seed Fingerprints according to ZIP 32

hardened_only

Generic framework for hardened-only key derivation.

Structs

AccountId

A type-safe wrapper for account identifiers.

ChainCode

A value that is needed, in addition to a spending key, in order to derive descendant keys and addresses of that key.

ChildIndex

A child index for a derived key.

DiversifierIndex

The index for a particular diversifier.

DiversifierIndexOverflowError

The error type returned when a DiversifierIndex increment fails.

TryFromIntError

The error type returned when a checked integral type conversion fails.

Enums

Scope

The scope of a viewing key or address.