zcash_primitives::zip32::registered

Function cryptovalue_from_subpath

pub fn cryptovalue_from_subpath(
    context_string: &[u8],
    seed: &[u8],
    zip_number: u16,
    subpath: &[PathElement<'_>],
) -> Result<[u8; 64], DerivationError>

Derives a 64-byte cryptovalue (for use as key material for example), for a registered application protocol at the given non-empty subpath from the given seed. Each subpath element may consist of an index and a (possibly empty) tag.

context_string: an identifier for the context in which this key will be used. It must be globally unique, non-empty, and no more than 252 bytes in length.

seed: the root seed. Must be between 32 bytes and 252 bytes in length, inclusive.

zip_number: the number of the ZIP defining the application protocol. The corresponding hardened index (with empty tag) will be prepended to the subpath to obtain the ZIP 32 path.

subpath: the path to the desired child element. A non-empty path is required, in order to ensure that the resulting full-width cryptovalue is within the allowed subtree rooted at m_{context} / zip_number'.