Crate zcash_primitives

General Zcash primitives.

zcash_primitives is a library that provides the core structs and functions necessary for working with Zcash.

Feature flags

circuits (enabled by default) — Enables creating proofs

multicore (enabled by default) — Enables multithreading support for creating proofs.

transparent-inputs — Enables spending transparent notes with the transaction builder.

test-dependencies — Exposes APIs that are useful for testing, such as proptest strategies.

expensive-tests — A feature used to isolate tests that are expensive to run. Test-only.

no-expensive-tests — A feature used to ensure that expensive tests do not run when running with –all-features

non-standard-fees — A feature that provides FeeRule implementations and constructors for non-standard fees.

Modules

block

Structs and methods for handling Zcash block headers.

merkle_tree

Parsers and serializers for Zcash Merkle trees.

transaction

Structs and methods for handling Zcash transactions.