Crate zcash_primitives

General Zcash primitives.

zcash_primitives is a library that provides the core structs and functions necessary for working with Zcash.

Modules

block

Structs and methods for handling Zcash block headers.

consensus

Consensus logic and parameters.

constants

Various constants used by the Zcash primitives.

extensions

keys

legacy

Support for legacy transparent addresses and scripts.

memo

Structs for handling encrypted memos.

merkle_tree

Implementation of a Merkle tree of commitments used to prove the existence of notes.

sapling

Structs and constants specific to the Sapling shielded pool.

transaction

Structs and methods for handling Zcash transactions.

zip32

Implementation of ZIP 32 for hierarchical deterministic key management.

zip339

ZIP 339 for seed phrase support.