Crate zcash_primitives

General Zcash primitives.

zcash_primitives is a library that provides the core structs and functions necessary for working with Zcash.

Feature flags

multicore (enabled by default) — Enables multithreading support for creating proofs.

transparent-inputs — Enables spending transparent notes with the transaction builder.

test-dependencies — Exposes APIs that are useful for testing, such as proptest strategies.

expensive-tests — A feature used to isolate tests that are expensive to run. Test-only.

non-standard-fees — A feature that provides FeeRule implementations and constructors for non-standard fees.

Re-exports

pub use zip32;

Modules

block

Structs and methods for handling Zcash block headers.

consensus

Consensus logic and parameters.

constants

Network-specific Zcash constants.

legacy

Support for legacy transparent addresses and scripts.

memo

Structs for handling encrypted memos.

merkle_tree

Parsers and serializers for Zcash Merkle trees.

transaction

Structs and methods for handling Zcash transactions.