Function zcash_primitives::sapling::note_encryption::sapling_note_encryption

pub fn sapling_note_encryption<R: RngCore, P: Parameters>(
    ovk: Option<OutgoingViewingKey>, 
    note: Note, 
    to: PaymentAddress, 
    memo: MemoBytes, 
    rng: &mut R
) -> NoteEncryption<SaplingDomain<P>>

Creates a new encryption context for the given note.

Setting ovk to None represents the ovk = ⊥ case, where the note cannot be recovered by the sender.