zcash_primitives::transaction
Module builder

Structs for building transactions.

Structs

BuildResult

The result of a transaction build operation, which includes the resulting transaction along with metadata describing how spends and outputs were shuffled in creating the transaction’s shielded bundles.

Builder

Generates a Transaction from its inputs and outputs.

PcztParts

The components of a PCZT.

PcztResult

The result of Builder::build_for_pczt.

Progress

Reports on the progress made by the builder towards building a transaction.

Enums

BuildConfig

Rules for how the builder should be configured for each shielded pool.

Error

Errors that can occur during transaction construction.

FeeError

Errors that can occur during fee calculation.

Constants

DEFAULT_TX_EXPIRY_DELTA

Since Blossom activation, the default transaction expiry delta should be 40 blocks. https://zips.z.cash/zip-0203#changes-for-blossom