zcash_primitives::transaction::builder

Struct Builder

pub struct Builder<'a, P, U: ProverProgress> { /* private fields */ }

Generates a Transaction from its inputs and outputs.

Implementations

impl<P: Parameters, U: ProverProgress> Builder<'_, P, U>

pub fn mock_build<R: RngCore>( self, transparent_signing_set: &TransparentSigningSet, sapling_extsks: &[ExtendedSpendingKey], orchard_saks: &[SpendAuthorizingKey], rng: R, ) -> Result<BuildResult, Error<FeeError>>

Available on crate feature test-dependencies only.

Build the transaction using mocked randomness and proving capabilities. DO NOT USE EXCEPT FOR UNIT TESTING.

impl<P, U: ProverProgress> Builder<'_, P, U>

pub fn params(&self) -> &P

Returns the network parameters that the builder has been configured for.

pub fn target_height(&self) -> BlockHeight

Returns the target height of the transaction under construction.

pub fn transparent_inputs(&self) -> &[TransparentInputInfo]

Available on crate feature transparent-inputs only.

Returns the set of transparent inputs currently committed to be consumed by the transaction.

pub fn transparent_outputs(&self) -> &[TxOut]

Returns the set of transparent outputs currently set to be produced by the transaction.

pub fn sapling_inputs(&self) -> &[SpendInfo]

Returns the set of Sapling inputs currently committed to be consumed by the transaction.

pub fn sapling_outputs(&self) -> &[OutputInfo]

Returns the set of Sapling outputs currently set to be produced by the transaction.

impl<'a, P: Parameters> Builder<'a, P, ()>

pub fn new( params: P, target_height: BlockHeight, build_config: BuildConfig, ) -> Self

Creates a new Builder targeted for inclusion in the block with the given height, using default values for general transaction fields.

Default values

The expiry height will be set to the given height plus the default transaction expiry delta (20 blocks).

pub fn with_progress_notifier( self, _progress_notifier: Sender<Progress>, ) -> Builder<'a, P, Sender<Progress>>

Available on crate feature std only.

Sets the notifier channel, where progress of building the transaction is sent.

An update is sent after every Sapling Spend or Output is computed, and the u32 sent represents the total steps completed so far. It will eventually send number of spends + outputs. If there’s an error building the transaction, the channel is closed.

impl<P: Parameters, U: ProverProgress> Builder<'_, P, U>

pub fn add_orchard_spend<FE>( &mut self, fvk: FullViewingKey, note: Note, merkle_path: MerklePath, ) -> Result<(), Error<FE>>

Adds an Orchard note to be spent in this bundle.

Returns an error if the given Merkle path does not have the required anchor for the given note.

pub fn add_orchard_output<FE>( &mut self, ovk: Option<OutgoingViewingKey>, recipient: Address, value: u64, memo: MemoBytes, ) -> Result<(), Error<FE>>

Adds an Orchard recipient to the transaction.

pub fn add_sapling_spend<FE>( &mut self, fvk: FullViewingKey, note: Note, merkle_path: MerklePath, ) -> Result<(), Error<FE>>

Adds a Sapling note to be spent in this transaction.

Returns an error if the given Merkle path does not have the same anchor as the paths for previous Sapling notes.

pub fn add_sapling_output<FE>( &mut self, ovk: Option<OutgoingViewingKey>, to: PaymentAddress, value: Zatoshis, memo: MemoBytes, ) -> Result<(), Error<FE>>

Adds a Sapling address to send funds to.

pub fn add_transparent_input( &mut self, pubkey: PublicKey, utxo: OutPoint, coin: TxOut, ) -> Result<(), Error>

Available on crate feature transparent-inputs only.

Adds a transparent coin to be spent in this transaction.

pub fn add_transparent_output( &mut self, to: &TransparentAddress, value: Zatoshis, ) -> Result<(), Error>

Adds a transparent address to send funds to.

pub fn get_fee<FR: FeeRule>( &self, fee_rule: &FR, ) -> Result<Zatoshis, FeeError<FR::Error>>

Reports the calculated fee given the specified fee rule.

This fee is a function of the spends and outputs that have been added to the builder, pursuant to the specified FeeRule.

pub fn build<R: RngCore + CryptoRng, SP: SpendProver, OP: OutputProver, FR: FeeRule>( self, transparent_signing_set: &TransparentSigningSet, sapling_extsks: &[ExtendedSpendingKey], orchard_saks: &[SpendAuthorizingKey], rng: R, spend_prover: &SP, output_prover: &OP, fee_rule: &FR, ) -> Result<BuildResult, Error<FR::Error>>

Available on crate feature circuits only.

Builds a transaction from the configured spends and outputs.

Upon success, returns a tuple containing the final transaction, and the [SaplingMetadata] generated during the build process.

pub fn build_for_pczt<R: RngCore + CryptoRng, FR: FeeRule>( self, rng: R, fee_rule: &FR, ) -> Result<PcztResult<P>, Error<FR::Error>>

Builds a PCZT from the configured spends and outputs.

Upon success, returns a struct containing the PCZT components, and the [SaplingMetadata] and [orchard::builder::BundleMetadata] generated during the build process.