zcash_primitives::transaction::components::transparent

Module pczt

👎Deprecated: This module is deprecated; use ::zcash_transparent::pczt instead.

Structs

Bip32Derivation

The BIP 32 derivation path at which a key can be found.

Bundle

PCZT fields that are specific to producing the transaction’s transparent bundle (if any).

Input

Information about a transparent spend within a transaction.

InputUpdater

An updater for a transparent PCZT input.

Output

Information about a transparent output within a transaction.

OutputUpdater

An updater for a transparent PCZT output.

Unbound

Authorizing data for a transparent bundle in a transaction that is just missing binding signatures.

Updater

An updater for a transparent PCZT bundle.

Enums

ParseError

Errors that can occur while parsing a PCZT bundle.

SignerError

Errors that can occur while signing a transparent input in a PCZT.

SpendFinalizerError

Errors that can occur while finalizing the transparent inputs of a PCZT bundle.

TxExtractorError

Errors that can occur while extracting a regular transparent bundle from a PCZT bundle.

UpdaterError

Errors that can occur while signing a transparent input in a PCZT.

VerifyError

Errors that can occur while verifying a PCZT bundle.